app/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0
chromadb==0.5.3
sentence-transformers==3.0.1
PyYAML==6.0.2
